requirements.txt
requests
pandas
ta
python-dotenv
schedule
python-telegram-bot==13.15
joblib
scikit-learn
xgboost